95dff4d fixing mac warnings
76c533c down to 41 warnings
7754f7c change the difference of pointers to ptrdiff_t
4309360 change the difference of pointers to ptrdiff_t
5fbe13b more warning fixes for win64
131dacf endianness swap done with streamoff, not int
ff00ae8 more warning fixes for win64
59e93cc more warning fixes, including static call for Unpacker12Bits
1d5a601 more warning fixes for win64
3fae58d more warning fixes for win64
106b167 more warning fixes for win64, including streamoff
591702a more warning fixes for win64
5c73a95 warning fixes in the anonymizer
d9b139d trying to fix things to add a comment
ca9a003 Another compile fix
556f4fb Another compile fix
a8d1eaf Remove another size_t warning
7aead56 Another compile fix
e900b11 Actual fix
d5a3693 Different type of fix for size_t on win64
b10b3d5 Fix a warning on win64 bit about size_T

+++ Utilities/gdcm/Source/Common/gdcmSwapper.h
@@ -48,7 +48,7 @@ class SwapperNoOp
 {
 public:
   template <typename T> static T Swap(T val) {return val;}
-  template <typename T> static void SwapArray(T *, unsigned int ) {}
+  template <typename T> static void SwapArray(T *, size_t ) {}
 };
 
 class SwapperDoOp
@@ -56,10 +56,10 @@ class SwapperDoOp
 public:
   template <typename T> static T Swap(T val);
   template <typename T>
-  static void SwapArray(T *array, unsigned int n)
+  static void SwapArray(T *array, size_t n)
   {
     // TODO: need to unroll loop:
-    for(unsigned int i = 0; i < n; ++i)
+    for(size_t i = 0; i < n; ++i)
     {
       array[i] = Swap<T>(array[i]);
     }
